Added Dionysis Zindros' gist for the WoT system (also depending on graph theory) which is to be applied to the OpenBazaar project.
This appears to be the best and most comprehensive real world example
of what would be a realm within this proposal.  That is, it is
entirely self-contained within the definitions of the OpenBazaar
software and protocol, but could be used in conjunction with a wider
WoT utilising either pseudonymous or "real identity" node identifiers
(via cryptographic verification).